requirements.txt
python-dotenv
flask_cors
flask
seaborn
pandas
plotly